move CertLogic JS/TS validator into certlogic-js
