+++ certlogic/build-js.sh
@@ -2,5 +2,3 @@
 
 (cd certlogic-js ; sh build.sh)
 
-(cd certlogic-validation ; sh build.sh)
-
