+++ certlogic/certlogic-js/src/validation/data-accesses.ts
@@ -1,4 +1,4 @@
-import { CertLogicExpression } from "certlogic-js"
+import { CertLogicExpression } from "../typings"
 
 
 const gatherFrom = (expr: CertLogicExpression): string[] => {
@@ -28,7 +28,7 @@ const gatherFrom = (expr: CertLogicExpression): string[] => {
         if (operator === "reduce") {
             return [ /* operand: */...gatherFrom(values[0]), /* initial: */...gatherFrom(values[2]) ]
         }
-        throw new Error(`operator not recognised by fields gatherer ("gatherFields") in certlogic-validation/${__filename}: "${operator}"`)
+        throw new Error(`operator not recognised by fields gatherer ("gatherFields") in certlogic-js/validation/${__filename}: "${operator}"`)
     }
     return []
 }
