+++ certlogic/certlogic-js/tsconfig.json
@@ -3,10 +3,10 @@
     "declaration": true,
     "declarationMap": true,
     "strict": true,
-    "target": "ES2015",
+    "target": "ES2020",
     "sourceMap": true,
     "outDir": "dist",
-    "lib": [ "ES2015" ],
+    "lib": [ "ES2020" ],
     "module": "commonjs"
   },
   "include": [
