--- certlogic/certlogic-validation/.gitignore
@@ -1,5 +1,0 @@
-/dist
-/node_modules
-/package-lock.json
-/yarn.lock
-/yarn-error.log
