--- certlogic/certlogic-validation/build.sh
@@ -1,6 +1,0 @@
-#!/usr/bin/env sh
-
-npm run clean
-npm install
-npm test
-
