--- certlogic/certlogic-validation/package.json
@@ -1,49 +1,0 @@
-{
-  "name": "certlogic-validation",
-  "version": "0.8.1",
-  "description": "Validators for CertLogic",
-  "keywords": [
-    "json",
-    "logic",
-    "jsonlogic",
-    "rules",
-    "validation",
-    "validator"
-  ],
-  "homepage": "https://github.com/ehn-dcc-development/dgc-business-rules/tree/main/certlogic",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/ehn-dcc-development/dgc-business-rules.git"
-  },
-  "bugs": {
-    "url": "https://github.com/ehn-dcc-development/dgc-business-rules/issues"
-  },
-  "main": "dist/index.js",
-  "scripts": {
-    "build": "tsc",
-    "build-watch": "tsc --watch --incremental",
-    "pretest": "npm run build",
-    "test": "mocha dist/test",
-    "test-watch": "mocha --watch dist/test",
-    "clean": "rm -rf dist/ && rm -rf node_modules/ && rm -rf package-lock.json && rm -rf yarn.lock"
-  },
-  "bin": {
-    "certlogic-validate": "dist/cli.js"
-  },
-  "author": "Meinte Boersma <meinte.boersma@gmail.com>",
-  "license": "Apache-2.0",
-  "dependencies": {
-    "@types/json-schema": "^7.0.7",
-    "certlogic-js": "0.8.0",
-    "deep-equal": "^2.0.5"
-  },
-  "devDependencies": {
-    "@types/chai": "^4.2.18",
-    "@types/deep-equal": "^1.0.1",
-    "@types/mocha": "^8.2.2",
-    "@types/node": "^15.12.1",
-    "chai": "^4.3.4",
-    "mocha": "^8.4.0",
-    "typescript": "^4.3.4"
-  }
-}
