--- certlogic/certlogic-validation/tsconfig.json
@@ -1,19 +1,0 @@
-{
-  "compilerOptions": {
-    "moduleResolution": "Node",
-    "declaration": true,
-    "declarationMap": true,
-    "strict": true,
-    "target": "ES2020",
-    "sourceMap": true,
-    "outDir": "dist",
-    "lib": [ "ES2020" ],
-    "module": "commonjs"
-  },
-  "include": [
-    "src/**/*.ts"
-  ],
-  "exclude": [
-    "node_modules"
-  ]
-}
