+++ jsonlogic/javascript/package.json
@@ -16,7 +16,7 @@
     "@types/json-logic-js": "^1.2.1",
     "@types/mocha": "^8.2.2",
     "@types/node": "^15.12.2",
-    "certlogic-validation": "^0.8.0",
+    "certlogic-js": "^0.8.2",
     "chai": "^4.3.4",
     "deep-equal": "^2.0.5",
     "mocha": "^9.0.1"
